[scheduling] implement backtracking of intermediate results
For batch programs, we currently schedule all tasks which are sources
and let them kick off the execution of the connected tasks. This
approach bears some problems when executing large dataflows with many
branches. With backtracking, we traverse the execution graph
output-centrically (from the sinks) in a depth-first manner. This
enables us to use resources differently.

In the course of backtracking, only tasks will be executed that are
required to supply inputs to the current task. When a job is newly
submitted, this means that the backtracking will reach the
sources. When the job has been previously executed and intermediate
results are available, old ResultPartitions to resume from can be
requested while backtracking.

Backtracking is disabled by default. It can be enabled by setting the
ScheduleMode in JobGraph to BACKTRACKING.

CHANGELOG
- new scheduling mode: backtracking
- backtracks from the sinks of an ExecutionGraph
- checks the availability of IntermediatePartitionResults
- marks ExecutionVertex to be scheduled
- caches ResultPartitions and reloads them
- resumes from intermediate results
- test for general behavior of backtracking (BacktrackingTest)
- test for resuming from an intermediate result (ResumeITCase)
- test for releasing of cached ResultPartitions (ResultPartitionManagerTest)
- allow multiple consumers per blocking intermediate result (batch)